Add EnumStoreBuilder
