+++ src/libraries/Microsoft.PowerFx.Core/Glue/SimpleResolver.cs
@@ -9,7 +9,6 @@ using Microsoft.PowerFx.Core.Binding;
 using Microsoft.PowerFx.Core.Binding.BindInfo;
 using Microsoft.PowerFx.Core.Entities;
 using Microsoft.PowerFx.Core.Functions;
-using Microsoft.PowerFx.Core.Texl;
 using Microsoft.PowerFx.Core.Types;
 using Microsoft.PowerFx.Core.Types.Enums;
 using Microsoft.PowerFx.Core.Utils;
@@ -60,7 +59,7 @@ namespace Microsoft.PowerFx.Core.Glue
             _config = config ?? throw new ArgumentNullException(nameof(config));
 
             _library = config.Functions.ToArray();
-            _enums = config.EnumStore.WithRequiredEnums(_library).EnumSymbols.ToArray();
+            _enums = config.EnumStoreBuilder.Build().EnumSymbols.ToArray();
         }
 
         // for derived classes that need to set INameResolver.Document.
